Adding the original import script back into Wrapper: Offline for backwards compatibility
(also in case people aren't able to get the new importer running for any reason and/or are on 32-bit machines)
I have also took the case of adding a restart feature when it reaches the original end of the script. I will also look forward to fixing the remaining bugs that it has.