Adding my modded export.bat file
So I finally fixed the bug involving the %ADDITIONAL% label not being recognised correctly by FFMPEG, after a bit of tinkering. I removed the set values for CRF and ADDITIONAL for all file types except mp4 because I was told that CRF is only supported by mp4 files - not to mention, if I left those variables in the code, the rendering process fails when exporting in mp4 but not AVI or WEBM.

The error is: "Unknown encoder 'aac -crf 17'". It will always display this error as "crf 17" even if a new CRF value is set. My modded export.bat file doesn't have this issue at all, and it is set up so that FFMPEG doesn't read the CRF value when it exports files in a AVI, WEBM or WMV format. For mp4, it will and will read the correct value specified by me ranging from 0 to 51 as intended. Anything higher or lower and it'll choose the nearest available number. (example: choosing a CRF of 78 causes it to go to 51)

However, exporting in WMV will need some work because all attempts to do so will fail due to a high sample rate. The error for that will always display as: "sample rate is too high: 192000 > 48kHz
Error initializing output stream 0:1 -- Error while opening encoder for output stream #0:1 - maybe incorrect parameters such as bit_rate, rate, width or height"

Hope that clears everything up lol