Modified the text in the LVM
This was done mainly to reduce the instances of "GoAnimate" in the video maker and to get rid of or change some misleading text (also to show that we're not affiliated with GoAnimate or GoAnimate, Inc. at all). This applies to all video complexity warnings (150/200 scenes, 12+ minutes and very close zooming)

I also changed the message that appears when moving your mouse over the Save and Close button.

Before: "Save your animation and email it to your friends. This will close the studio."
After: "Save your animation and go to the video list, closing the studio."